+++ MatrixSDK/Data/MXRoomSummary.m
@@ -480,6 +480,30 @@ static NSUInteger const kMXRoomSummaryTrustComputationDelayMs = 1000;
 
 #pragma mark - Trust management
 
+- (void)enableTrustTracking:(BOOL)enable
+{
+    if (enable)
+    {
+        if (!_isEncrypted || _trust)
+        {
+            // Not applicable or already enabled
+            return;
+        }
+        
+        NSLog(@"[MXRoomSummary] enableTrustTracking: YES");
+        
+        // Bootstrap trust computation
+        [self registerTrustLevelDidChangeNotifications];
+        [self triggerComputeTrust:YES];
+    }
+    else
+    {
+        NSLog(@"[MXRoomSummary] enableTrustTracking: NO");
+        [self unregisterTrustLevelDidChangeNotifications];
+        _trust = nil;
+    }
+}
+
 - (void)setIsEncrypted:(BOOL)isEncrypted
 {
     _isEncrypted = isEncrypted;
@@ -516,6 +540,12 @@ static NSUInteger const kMXRoomSummaryTrustComputationDelayMs = 1000;
 {
     [[NSNotificationCenter defaultCenter] addObserver:self selector:@selector(deviceInfoTrustLevelDidChange:) name:MXDeviceInfoTrustLevelDidChangeNotification object:nil];
     [[NSNotificationCenter defaultCenter] addObserver:self selector:@selector(crossSigningInfoTrustLevelDidChange:) name:MXCrossSigningInfoTrustLevelDidChangeNotification object:nil];
+}
+
+- (void)unregisterTrustLevelDidChangeNotifications
+{
+    [[NSNotificationCenter defaultCenter] removeObserver:self name:MXDeviceInfoTrustLevelDidChangeNotification object:nil];
+    [[NSNotificationCenter defaultCenter] removeObserver:self name:MXCrossSigningInfoTrustLevelDidChangeNotification object:nil];
 }
 
 - (void)deviceInfoTrustLevelDidChange:(NSNotification*)notification
@@ -858,7 +888,11 @@ static NSUInteger const kMXRoomSummaryTrustComputationDelayMs = 1000;
         
         _hiddenFromUser = [aDecoder decodeBoolForKey:@"hiddenFromUser"];
         
-        if (_isEncrypted && [MXSDKOptions sharedInstance].computeE2ERoomSummaryTrust)
+        // Compute the trust if asked to do it automatically
+        // or maintain its computation it has been already calcutated
+        if (_isEncrypted
+            && ([MXSDKOptions sharedInstance].computeE2ERoomSummaryTrust
+                || _trust))
         {
             dispatch_async(dispatch_get_main_queue(), ^{
                 [self bootstrapTrustLevelComputation];